Add slivers to details view and the image carousel
